core: Fix race between AutoStartVmsRunner and ProcessDownVmCommand
Issue for killed HA VM:
  AutoStartVmsRunner acquires lock and runs the HA VM if there's
  no NEXT_RUN. It causes race condition between removing the NEXT_RUN
  snapshot in ProcessDownVmCommand.applyNextRunConfiguration() and
  calling UpdateVmCommand to apply it.

  The UpdateVmCommand fails to acquire locks if the RunVm command
  is executed right after the NEXT_RUN snapshot is removed.

Fixed by:
  ProcessDownVmCommand.applyNextRunConfiguration() acquires
  lock before removing the NEXT_RUN snapshot and passes it to
  the UpdateVmCommand.

  The AutoStartVmsRunner keeps failing to acquire the VM lock until
  the NEXT_RUN is applied.

Bug-Url: https://bugzilla.redhat.com/1369521

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/ProcessDownVmCommand.java
@@ -35,6 +35,8 @@ import org.ovirt.engine.core.common.errors.EngineMessage;
 import org.ovirt.engine.core.common.utils.VmDeviceCommonUtils;
 import org.ovirt.engine.core.common.utils.VmDeviceType;
 import org.ovirt.engine.core.compat.Guid;
+import org.ovirt.engine.core.utils.lock.EngineLock;
+import org.ovirt.engine.core.utils.lock.LockManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -53,6 +55,9 @@ public class ProcessDownVmCommand<T extends ProcessDownVmParameters> extends Com
 
     @Inject
     private SnapshotsManager snapshotsManager;
+
+    @Inject
+    private LockManager lockManager;
 
     private VmPool vmPoolCached;
 
@@ -227,18 +232,32 @@ public class ProcessDownVmCommand<T extends ProcessDownVmParameters> extends Com
         // and any newer update should be preffered to this one.
         Snapshot runSnap = snapshotDao.get(getVmId(), SnapshotType.NEXT_RUN);
         if (runSnap != null) {
-            snapshotDao.remove(runSnap.getId());
-            Date originalCreationDate = getVm().getVmCreationDate();
-            snapshotsManager.updateVmFromConfiguration(getVm(), runSnap.getVmConfiguration());
-            // override creation date because the value in the config is the creation date of the config, not the vm
-            getVm().setVmCreationDate(originalCreationDate);
+            log.debug("Attempt to apply NEXT_RUN snapshot for VM '{}'", getVmId());
 
-            VdcReturnValueBase result = runInternalAction(VdcActionType.UpdateVm, createUpdateVmParameters());
-            if (result.getActionReturnValue() != null && result.getActionReturnValue()
-                    .equals(VdcActionType.UpdateVmVersion)) { // Template-version changed
-                templateVersionChanged = true;
+            EngineLock updateVmLock = createUpdateVmLock();
+            if (lockManager.acquireLock(updateVmLock).getFirst()) {
+                snapshotDao.remove(runSnap.getId());
+                Date originalCreationDate = getVm().getVmCreationDate();
+                snapshotsManager.updateVmFromConfiguration(getVm(), runSnap.getVmConfiguration());
+                // override creation date because the value in the config is the creation date of the config, not the vm
+                getVm().setVmCreationDate(originalCreationDate);
+
+                VdcReturnValueBase result = runInternalAction(VdcActionType.UpdateVm, createUpdateVmParameters(),
+                        ExecutionHandler.createInternalJobContext(updateVmLock));
+                if (result.getActionReturnValue() != null && result.getActionReturnValue()
+                        .equals(VdcActionType.UpdateVmVersion)) { // Template-version changed
+                    templateVersionChanged = true;
+                }
+            } else {
+                log.warn("Could not acquire lock for UpdateVmCommand to apply Next Run Config of VM '{}'", getVmId());
             }
         }
+    }
+
+    private EngineLock createUpdateVmLock() {
+        return new EngineLock(
+                UpdateVmCommand.getExclusiveLocksForUpdateVm(getVm()),
+                UpdateVmCommand.getSharedLocksForUpdateVm(getVm()));
     }
 
     private VmManagementParametersBase createUpdateVmParameters() {
